refactor: extract execution-engine.ts from run.ts (#447)
Step 5: Extract executeWithClaude, executeWithProvider, and all execution
helpers (worktree management, foreground/background/watch modes, auto-commit,
verification, preflight checks) into dedicated module. -769 lines from run.ts.
