+++ src/commands/run.ts
@@ -1,7 +1,6 @@
 import ora from 'ora';
-import { spawn, execSync } from 'child_process';
-import { join, dirname } from 'path';
-import { existsSync, readFileSync, writeFileSync, mkdirSync, cpSync, unlinkSync } from 'fs';
+import { join } from 'path';
+import { existsSync, readFileSync } from 'fs';
 import {
   findSquadsDir,
   loadSquad,
@@ -10,8 +9,6 @@ import {
   parseAgentProvider,
   listSquads,
   findSimilarSquads,
-  EffortLevel,
-  Squad,
 } from '../lib/squad-parser.js';
 import {
   type RunOptions,
@@ -21,12 +18,7 @@ import {
 } from '../lib/run-types.js';
 import {
   generateExecutionId,
-  selectMcpConfig,
   detectTaskType,
-  type ClaudeModelAlias,
-  getClaudeModelAlias,
-  resolveModel,
-  ensureProjectTrusted,
   getProjectRoot,
   formatDuration,
   checkClaudeCliAvailable,
@@ -37,18 +29,12 @@ import {
   formatViolations,
   ExecutionRequest
 } from '../lib/permissions.js';
-import { findMemoryDir } from '../lib/memory.js';
 import {
-  type PreflightResult,
-  type Learning,
-  type ExecutionRecord,
   DEFAULT_SCHEDULED_COOLDOWN_MS,
-  registerContextWithBridge,
   checkPreflightGates,
   fetchLearnings,
   logExecution,
   updateExecutionStatus,
-  getLastExecutionTime,
   checkLocalCooldown,
   emitExecutionEvent,
 } from '../lib/execution-log.js';
@@ -66,13 +52,16 @@ import {
   getCLIConfig,
   isProviderCLIAvailable,
 } from '../lib/llm-clis.js';
-import { detectProviderFromModel } from '../lib/providers.js';
-import { loadSession, isLoggedIn } from '../lib/auth.js';
-import { getApiUrl, getBridgeUrl } from '../lib/env-config.js';
 import { runCloudDispatch } from '../lib/cloud-dispatch.js';
 import { runConversation, saveTranscript, type ConversationOptions } from '../lib/workflow.js';
 import { reportExecutionStart, reportConversationResult, pushCognitionSignal } from '../lib/api-client.js';
-import { getBotGitEnv, getBotPushUrl, getBotGhEnv, getCoAuthorTrailer } from '../lib/github.js';
+import { getBotGhEnv } from '../lib/github.js';
+import {
+  verifyExecution,
+  preflightExecutorCheck,
+  executeWithClaude,
+  executeWithProvider,
+} from '../lib/execution-engine.js';
 import {
   type LoopState,
   loadLoopState,
@@ -104,171 +93,18 @@ import {
 import { classifyAgent } from '../lib/conversation.js';
 
 // ── Operational constants (no magic numbers) ──────────────────────────
-const VERIFICATION_STATE_MAX_CHARS = 2000;
-const VERIFICATION_EXEC_TIMEOUT_MS = 30000;
 const DRYRUN_DEF_MAX_CHARS = 500;
 const DRYRUN_CONTEXT_MAX_CHARS = 800;
-const LOG_FILE_INIT_DELAY_MS = 500;
-const VERBOSE_COMMAND_MAX_CHARS = 50;
 
 // registerContextWithBridge, checkPreflightGates, fetchLearnings, logExecution,
 // updateExecutionStatus, getLastExecutionTime, checkLocalCooldown, emitExecutionEvent
 // → moved to src/lib/execution-log.ts
 
-/**
- * Auto-commit agent work after execution completes.
- * Commits as the Agents Squads bot (if configured), pushes with bot token.
- * Falls back to user's git identity if bot not configured.
- */
-async function autoCommitAgentWork(
-  squadName: string,
-  agentName: string,
-  executionId: string,
-  provider?: string,
-): Promise<{ committed: boolean; message?: string; error?: string }> {
-  const { execSync } = await import('child_process');
-  const { detectGitHubRepo } = await import('../lib/github.js');
-  const projectRoot = getProjectRoot();
-
-  try {
-    // Check for uncommitted changes
-    const status = execSync('git status --porcelain', {
-      encoding: 'utf-8',
-      cwd: projectRoot,
-    }).trim();
-
-    if (!status) {
-      return { committed: false };
-    }
-
-    // Get bot identity for commits
-    const botEnv = await getBotGitEnv();
-    const execOpts = {
-      cwd: projectRoot,
-      env: { ...process.env, ...botEnv },
-    };
-
-    // Stage all changes (agent work should be committed)
-    execSync('git add -A', execOpts);
-
-    // Build commit message with provider-specific co-author
-    // Write to temp file to avoid shell injection via squad/agent names
-    const shortExecId = executionId.slice(0, 12);
-    const coAuthor = getCoAuthorTrailer(provider || 'claude');
-    const msgFile = join(projectRoot, '.git', 'SQUADS_COMMIT_MSG');
-    writeFileSync(msgFile, `feat(${squadName}/${agentName}): execution ${shortExecId}\n\n${coAuthor}\n`);
-
-    // Commit using --file to avoid shell interpolation
-    try {
-      execSync(`git commit --file "${msgFile}"`, execOpts);
-    } finally {
-      try { unlinkSync(msgFile); } catch { /* ignore */ }
-    }
-
-    // Push to origin using bot token
-    try {
-      const { spawnSync } = await import('child_process');
-      const repo = detectGitHubRepo(projectRoot);
-      // Validate repo format (org/name) to prevent injection
-      if (repo && /^[\w.-]+\/[\w.-]+$/.test(repo)) {
-        const pushUrl = await getBotPushUrl(repo);
-        if (pushUrl) {
-          // Use spawnSync with args array to avoid shell injection
-          spawnSync('git', ['push', pushUrl, 'HEAD'], { ...execOpts, stdio: 'pipe' });
-        } else {
-          spawnSync('git', ['push', 'origin', 'HEAD'], { ...execOpts, stdio: 'pipe' });
-        }
-      } else {
-        spawnSync('git', ['push', 'origin', 'HEAD'], { ...execOpts, stdio: 'pipe' });
-      }
-    } catch (e) {
-      writeLine(`  ${colors.dim}warn: git push failed (commit is still local): ${e instanceof Error ? e.message : String(e)}${RESET}`);
-    }
-
-    return { committed: true, message: `Committed changes from ${agentName}` };
-  } catch (error) {
-    return { committed: false, error: String(error) };
-  }
-}
-
-/**
- * Verify execution against acceptance criteria using a lightweight model.
- * Returns pass/fail with reason. Used by the Ralph verification loop.
- */
-async function verifyExecution(
-  squadName: string,
-  agentName: string,
-  criteria: string,
-  options: { verbose?: boolean } = {}
-): Promise<{ passed: boolean; reason: string }> {
-  const { execSync } = await import('child_process');
-  const projectRoot = getProjectRoot();
-
-  // Gather evidence: state file + recent commits
-  let stateContent = '';
-  const memDir = findMemoryDir();
-  if (memDir) {
-    const statePath = join(memDir, squadName, agentName, 'state.md');
-    if (existsSync(statePath)) {
-      stateContent = readFileSync(statePath, 'utf-8').slice(0, VERIFICATION_STATE_MAX_CHARS);
-    }
-  }
-
-  let recentCommits = '';
-  try {
-    recentCommits = execSync('git log --oneline -5 --no-color', {
-      encoding: 'utf-8',
-      cwd: projectRoot,
-    }).trim();
-  } catch (e) {
-    if (options.verbose) writeLine(`  ${colors.dim}warn: git log failed: ${e instanceof Error ? e.message : String(e)}${RESET}`);
-    recentCommits = '(no commits found)';
-  }
-
-  const verifyPrompt = `You are verifying whether an agent completed its task successfully.
-
-Agent: ${squadName}/${agentName}
-
-## Acceptance Criteria
-${criteria}
-
-## Evidence
-
-### Agent State File
-${stateContent || '(empty or not found)'}
-
-### Recent Git Commits
-${recentCommits}
-
-## Instructions
-Evaluate whether the acceptance criteria are met based on the evidence.
-Respond with EXACTLY one line:
-PASS: <brief reason>
-or
-FAIL: <brief reason>`;
-
-  try {
-    const escapedPrompt = verifyPrompt.replace(/'/g, "'\\''");
-    const result = execSync(
-      `unset CLAUDECODE; claude --print --model haiku -- '${escapedPrompt}'`,
-      { encoding: 'utf-8', cwd: projectRoot, timeout: VERIFICATION_EXEC_TIMEOUT_MS, shell: '/bin/sh' }
-    ).trim();
-
-    if (options.verbose) {
-      writeLine(`  ${colors.dim}Verification: ${result}${RESET}`);
-    }
-
-    if (result.startsWith('PASS')) {
-      return { passed: true, reason: result.replace(/^PASS:\s*/, '') };
-    }
-    return { passed: false, reason: result.replace(/^FAIL:\s*/, '') };
-  } catch (error) {
-    if (options.verbose) {
-      writeLine(`  ${colors.dim}Verification error (defaulting to PASS): ${error}${RESET}`);
-    }
-    return { passed: true, reason: 'Verification unavailable — defaulting to pass' };
-  }
-}
+// autoCommitAgentWork, verifyExecution, preflightExecutorCheck, executeWithClaude,
+// executeWithProvider, buildAgentEnv, logVerboseExecution, resolveTargetRepoRoot,
+// createAgentWorktree, cleanupWorktree, buildDetachedShellScript, prepareLogFiles,
+// executeForeground, executeWatch, ExecuteWithClaudeOptions
+// → moved to src/lib/execution-engine.ts
 
 // runCloudDispatch → moved to src/lib/cloud-dispatch.ts
 
@@ -1617,618 +1453,13 @@ TIME LIMIT: You have ${timeoutMins} minutes. Work efficiently:
   }
 }
 
-async function checkClaudeCliAvailable(): Promise<boolean> {
-  return new Promise((resolve) => {
-    const check = spawn('which', ['claude'], { stdio: 'pipe' });
-    check.on('close', (code) => resolve(code === 0));
-    check.on('error', () => resolve(false));
-  });
-}
-
-/**
- * Pre-flight check for the executor (Claude Code or other provider CLI).
- * Runs once at the start of `squads run` before any agent execution.
- * Checks:
- *   1. CLI binary is available on PATH
- *   2. Authentication looks configured (credentials file or API key)
- * Skippable with SQUADS_SKIP_CHECKS=1 env var (for CI/CD).
- * Returns true if checks pass (or are skipped), false if execution should abort.
- */
-async function preflightExecutorCheck(provider: string): Promise<boolean> {
-  // Allow skipping for CI/CD or advanced users
-  if (process.env.SQUADS_SKIP_CHECKS === '1') {
-    return true;
-  }
-
-  const isAnthropic = provider === 'anthropic';
-
-  // --- Check 1: CLI binary on PATH ---
-  let cliFound: boolean;
-
-  if (isAnthropic) {
-    cliFound = await checkClaudeCliAvailable();
-  } else {
-    cliFound = isProviderCLIAvailable(provider);
-  }
-
-  if (!cliFound) {
-    const cliConfig = getCLIConfig(provider);
-    const cliName = cliConfig?.command || provider;
-    const installCmd = cliConfig?.install || `See ${provider} documentation`;
-
-    writeLine();
-    writeLine(`  ${icons.error} ${colors.red}${cliName} CLI not found${RESET}`);
-    writeLine();
-    writeLine(`  ${colors.dim}The ${cliName} command is required to run agents but was not found on your PATH.${RESET}`);
-    writeLine();
-    writeLine(`  ${colors.cyan}Install:${RESET} ${installCmd}`);
-    writeLine();
-    writeLine(`  ${colors.dim}Skip this check: SQUADS_SKIP_CHECKS=1 squads run ...${RESET}`);
-    writeLine();
-    return false;
-  }
-
-  // Auth check removed: Claude CLI handles its own auth errors with clear messages.
-  // Pre-checking here caused false warnings for OAuth users (keychain auth works
-  // without .credentials.json or ANTHROPIC_API_KEY). See #520.
-
-  return true;
-}
-
-interface ExecuteWithClaudeOptions {
-  verbose?: boolean;
-  timeoutMinutes?: number;
-  foreground?: boolean; // Deprecated, now default
-  background?: boolean; // Opt-in background mode
-  watch?: boolean; // Background but tail log
-  useApi?: boolean;
-  effort?: EffortLevel;
-  skills?: string[];
-  trigger?: ExecutionContext['trigger'];
-  squadName: string;
-  agentName: string;
-  model?: string; // Model to use (Claude aliases or full model IDs like gemini-2.5-flash)
-}
-
-/** Build agent environment variables for Claude execution */
-function buildAgentEnv(
-  baseEnv: Record<string, string>,
-  execContext: ExecutionContext,
-  options?: { effort?: EffortLevel; skills?: string[]; includeOtel?: boolean; ghToken?: string }
-): Record<string, string> {
-  // Strip CLAUDECODE to allow spawning claude from within a Claude Code session
-  const { CLAUDECODE: _, ...cleanEnv } = baseEnv;
-  const env: Record<string, string> = {
-    ...cleanEnv,
-    SQUADS_SQUAD: execContext.squad,
-    SQUADS_AGENT: execContext.agent,
-    SQUADS_TASK_TYPE: execContext.taskType,
-    SQUADS_TRIGGER: execContext.trigger,
-    SQUADS_EXECUTION_ID: execContext.executionId,
-    BRIDGE_API: getBridgeUrl(),
-  };
-
-  // Inject bot GH_TOKEN so agents create PRs/issues as the bot identity,
-  // not the user's personal gh auth. This enables founder to review/approve.
-  if (options?.ghToken) env.GH_TOKEN = options.ghToken;
-
-  if (options?.includeOtel) {
-    env.OTEL_RESOURCE_ATTRIBUTES = `squads.squad=${execContext.squad},squads.agent=${execContext.agent},squads.task_type=${execContext.taskType},squads.trigger=${execContext.trigger},squads.execution_id=${execContext.executionId}`;
-  }
-
-  if (options?.effort) env.CLAUDE_EFFORT = options.effort;
-  if (options?.skills && options.skills.length > 0) env.CLAUDE_SKILLS = options.skills.join(',');
-
-  return env;
-}
-
-/** Log verbose execution config (shared by foreground and background modes) */
-function logVerboseExecution(config: {
-  projectRoot: string;
-  mode: string;
-  useApi?: boolean;
-  execContext: ExecutionContext;
-  effort?: EffortLevel;
-  skills?: string[];
-  resolvedModel?: string;
-  claudeModelAlias?: string;
-  explicitModel?: string;
-  logFile?: string;
-  mcpConfigPath?: string;
-}): void {
-  writeLine(`  ${colors.dim}Project: ${config.projectRoot}${RESET}`);
-  writeLine(`  ${colors.dim}Mode: ${config.mode}${RESET}`);
-  if (config.logFile) writeLine(`  ${colors.dim}Log: ${config.logFile}${RESET}`);
-  if (config.mcpConfigPath) writeLine(`  ${colors.dim}MCP config: ${config.mcpConfigPath}${RESET}`);
-  if (config.useApi !== undefined) writeLine(`  ${colors.dim}Auth: ${config.useApi ? 'API credits' : 'subscription'}${RESET}`);
-  writeLine(`  ${colors.dim}Execution: ${config.execContext.executionId}${RESET}`);
-  writeLine(`  ${colors.dim}Task type: ${config.execContext.taskType}${RESET}`);
-  writeLine(`  ${colors.dim}Trigger: ${config.execContext.trigger}${RESET}`);
-  if (config.effort) writeLine(`  ${colors.dim}Effort: ${config.effort}${RESET}`);
-  if (config.skills && config.skills.length > 0) writeLine(`  ${colors.dim}Skills: ${config.skills.join(', ')}${RESET}`);
-  if (config.resolvedModel || config.claudeModelAlias) {
-    const source = config.explicitModel ? 'explicit' : 'auto-routed';
-    const displayModel = config.resolvedModel || config.claudeModelAlias;
-    writeLine(`  ${colors.dim}Model: ${displayModel} (${source})${RESET}`);
-  }
-}
-
-/** Resolve the target repo root from the squad's repo field (e.g. "org/squads-cli" → sibling dir) */
-function resolveTargetRepoRoot(projectRoot: string, squad: Squad | null): string {
-  if (!squad?.repo) return projectRoot;
-  const repoName = squad.repo.split('/').pop();
-  if (!repoName) return projectRoot;
-  const candidatePath = join(projectRoot, '..', repoName);
-  return existsSync(candidatePath) ? candidatePath : projectRoot;
-}
-
-/** Create an isolated worktree for agent execution (Node.js-based, for foreground mode) */
-function createAgentWorktree(projectRoot: string, squadName: string, agentName: string): string {
-  const timestamp = Date.now();
-  const branchName = `agent/${squadName}/${agentName}-${timestamp}`;
-  const worktreePath = join(projectRoot, '..', '.worktrees', `${squadName}-${agentName}-${timestamp}`);
-
-  try {
-    mkdirSync(join(projectRoot, '..', '.worktrees'), { recursive: true });
-    execSync(`git worktree add '${worktreePath}' -b '${branchName}' HEAD`, { cwd: projectRoot, stdio: 'pipe' });
-    return worktreePath;
-  } catch (e) {
-    writeLine(`  ${colors.dim}warn: worktree creation failed, using project root: ${e instanceof Error ? e.message : String(e)}${RESET}`);
-    return projectRoot;
-  }
-}
-
-/** Remove a worktree and its branch after agent execution completes */
-function cleanupWorktree(worktreePath: string, projectRoot: string): void {
-  if (worktreePath === projectRoot) return; // fallback mode, nothing to clean
-
-  try {
-    // Extract branch name from worktree before removing
-    const branchInfo = execSync(`git -C '${projectRoot}' worktree list --porcelain`, { encoding: 'utf-8' });
-    let branchName = '';
-    const lines = branchInfo.split('\n');
-    for (let i = 0; i < lines.length; i++) {
-      if (lines[i] === `worktree ${worktreePath}` && i + 2 < lines.length) {
-        const branchLine = lines[i + 2]; // "branch refs/heads/..."
-        if (branchLine.startsWith('branch refs/heads/')) {
-          branchName = branchLine.replace('branch refs/heads/', '');
-        }
-        break;
-      }
-    }
-
-    // Remove worktree
-    execSync(`git -C '${projectRoot}' worktree remove '${worktreePath}' --force`, { stdio: 'pipe' });
-
-    // Delete the agent branch (only agent/* branches, safety check)
-    if (branchName && branchName.startsWith('agent/')) {
-      execSync(`git -C '${projectRoot}' branch -D '${branchName}'`, { stdio: 'pipe' });
-    }
-  } catch {
-    // Non-critical — worktree prune will catch it later
-  }
-}
-
-/** Build shell script for detached execution with worktree isolation */
-function buildDetachedShellScript(config: {
-  projectRoot: string;
-  squadName: string;
-  agentName: string;
-  timestamp: number;
-  claudeModelAlias?: string;
-  escapedPrompt: string;
-  logFile: string;
-  pidFile: string;
-}): string {
-  const modelFlag = config.claudeModelAlias ? `--model ${config.claudeModelAlias}` : '';
-  const branchName = `agent/${config.squadName}/${config.agentName}-${config.timestamp}`;
-  const worktreeDir = `${config.projectRoot}/../.worktrees/${config.squadName}-${config.agentName}-${config.timestamp}`;
-  const cleanup = `if [ "\${WORK_DIR}" != '${config.projectRoot}' ]; then git -C '${config.projectRoot}' worktree remove "\${WORK_DIR}" --force 2>/dev/null; BRANCH='${branchName}'; git -C '${config.projectRoot}' branch -D "\${BRANCH}" 2>/dev/null; fi`;
-  const script = `mkdir -p '${config.projectRoot}/../.worktrees'; WORK_DIR='${config.projectRoot}'; if git -C '${config.projectRoot}' worktree add '${worktreeDir}' -b '${branchName}' HEAD 2>/dev/null; then WORK_DIR='${worktreeDir}'; fi; cd "\${WORK_DIR}"; unset CLAUDECODE; claude --print --dangerously-skip-permissions --disable-slash-commands ${modelFlag} -- '${config.escapedPrompt}' > '${config.logFile}' 2>&1; ${cleanup}`;
-  return `echo $$ > '${config.pidFile}'; ${script}`;
-}
-
-/** Prepare log directory and file paths for detached execution */
-function prepareLogFiles(projectRoot: string, squadName: string, agentName: string, timestamp: number): { logDir: string; logFile: string; pidFile: string } {
-  const logDir = join(projectRoot, '.agents', 'logs', squadName);
-  const logFile = join(logDir, `${agentName}-${timestamp}.log`);
-  const pidFile = join(logDir, `${agentName}-${timestamp}.pid`);
-
-  if (!existsSync(logDir)) {
-    mkdirSync(logDir, { recursive: true });
-  }
-
-  return { logDir, logFile, pidFile };
-}
-
-/** Execute Claude in foreground mode (direct stdio, default) */
-function executeForeground(config: {
-  prompt: string;
-  claudeArgs: string[];
-  agentEnv: Record<string, string>;
-  projectRoot: string;
-  squadName: string;
-  agentName: string;
-  execContext: ExecutionContext;
-  startMs: number;
-  provider?: string;
-}): Promise<string> {
-  const workDir = createAgentWorktree(config.projectRoot, config.squadName, config.agentName);
-
-  return new Promise((resolve, reject) => {
-    const claude = spawn('claude', config.claudeArgs, {
-      stdio: 'inherit',
-      cwd: workDir,
-      env: config.agentEnv,
-    });
-
-    claude.on('close', async (code) => {
-      const durationMs = Date.now() - config.startMs;
-
-      if (code === 0) {
-        updateExecutionStatus(config.squadName, config.agentName, config.execContext.executionId, 'completed', {
-          outcome: 'Session completed successfully',
-          durationMs,
-        });
-
-        const commitResult = await autoCommitAgentWork(config.squadName, config.agentName, config.execContext.executionId, config.provider);
-        if (commitResult.committed) {
-          writeLine();
-          writeLine(`  ${colors.green}Auto-committed agent work${RESET}`);
-        }
-
-        cleanupWorktree(workDir, config.projectRoot);
-        resolve('Session completed');
-      } else {
-        updateExecutionStatus(config.squadName, config.agentName, config.execContext.executionId, 'failed', {
-          error: `Claude exited with code ${code}`,
-          durationMs,
-        });
-        cleanupWorktree(workDir, config.projectRoot);
-        reject(new Error(`Claude exited with code ${code}`));
-      }
-    });
-
-    claude.on('error', (err) => {
-      const durationMs = Date.now() - config.startMs;
-      updateExecutionStatus(config.squadName, config.agentName, config.execContext.executionId, 'failed', {
-        error: String(err),
-        durationMs,
-      });
-      cleanupWorktree(workDir, config.projectRoot);
-      reject(err);
-    });
-  });
-}
-
-/** Execute Claude in watch mode (background + tail log) */
-async function executeWatch(config: {
-  projectRoot: string;
-  agentEnv: Record<string, string>;
-  logFile: string;
-  wrapperScript: string;
-}): Promise<string> {
-  const child = spawn('sh', ['-c', config.wrapperScript], {
-    cwd: config.projectRoot,
-    detached: true,
-    stdio: 'ignore',
-    env: config.agentEnv,
-  });
-  child.unref();
-
-  await new Promise(resolve => setTimeout(resolve, LOG_FILE_INIT_DELAY_MS));
-
-  writeLine(`  ${colors.dim}Tailing log (Ctrl+C to stop watching, agent continues)...${RESET}`);
-  writeLine();
-
-  const tail = spawn('tail', ['-f', config.logFile], { stdio: 'inherit' });
-
-  process.on('SIGINT', () => {
-    tail.kill();
-    writeLine();
-    writeLine(`  ${colors.dim}Stopped watching. Agent continues in background.${RESET}`);
-    writeLine(`  ${colors.dim}Resume: tail -f ${config.logFile}${RESET}`);
-    process.exit(0);
-  });
-
-  return new Promise((resolve) => {
-    tail.on('close', () => {
-      resolve(`Agent running in background. Log: ${config.logFile}`);
-    });
-  });
-}
-
-async function executeWithClaude(
-  prompt: string,
-  options: ExecuteWithClaudeOptions
-): Promise<string> {
-  const {
-    verbose,
-    timeoutMinutes: _timeoutMinutes = 30,
-    foreground,
-    background,
-    watch,
-    useApi,
-    effort,
-    skills,
-    trigger = 'manual',
-    squadName,
-    agentName,
-    model,
-  } = options;
-
-  // Determine execution mode
-  const runInBackground = background === true && !watch;
-  const runInWatch = watch === true;
-  const runInForeground = !runInBackground && !runInWatch;
-
-  const startMs = Date.now();
-  const projectRoot = getProjectRoot();
-  ensureProjectTrusted(projectRoot);
-
-  // Resolve model and provider
-  const squad = squadName !== 'unknown' ? loadSquad(squadName) : null;
-  const mcpConfigPath = selectMcpConfig(squadName, squad);
-  const taskType = detectTaskType(agentName);
-  const resolvedModel = resolveModel(model, squad, taskType);
-  const provider = resolvedModel ? detectProviderFromModel(resolvedModel) : 'anthropic';
-
-  // Resolve target repo for worktree creation (squad.repo → sibling dir)
-  const targetRepoRoot = resolveTargetRepoRoot(projectRoot, squad);
-
-  // Delegate to non-Anthropic providers
-  if (provider !== 'anthropic' && provider !== 'unknown') {
-    if (verbose) {
-      const source = model ? 'explicit' : 'auto-routed';
-      writeLine(`  ${colors.dim}Model: ${resolvedModel} (${source})${RESET}`);
-      writeLine(`  ${colors.dim}Provider: ${provider}${RESET}`);
-    }
-    return executeWithProvider(provider, prompt, {
-      verbose, foreground, cwd: targetRepoRoot, squadName, agentName,
-    });
-  }
-
-  const claudeModelAlias = resolvedModel ? getClaudeModelAlias(resolvedModel) : undefined;
-
-  const execContext: ExecutionContext = {
-    squad: squadName, agent: agentName, taskType, trigger,
-    executionId: generateExecutionId(),
-  };
-
-  // Build base env: remove ANTHROPIC_API_KEY unless --use-api, remove CLAUDECODE
-  const { ANTHROPIC_API_KEY: _apiKey, CLAUDECODE: _claudeCode, ...envWithoutApiKey } = process.env;
-  const spawnEnv = useApi
-    ? (() => { const { CLAUDECODE: _, ...rest } = process.env; return rest; })()
-    : envWithoutApiKey;
-
-  const escapedPrompt = prompt.replace(/'/g, "'\\''");
-
-  await registerContextWithBridge(execContext);
-
-  // Get bot token so agents create PRs/issues as bot identity (not user's personal gh auth)
-  let botGhToken: string | undefined;
-  try {
-    const ghEnv = await getBotGhEnv();
-    botGhToken = ghEnv.GH_TOKEN;
-  } catch { /* graceful: falls back to user's gh auth */ }
-
-  // ── Foreground mode ──────────────────────────────────────────────────
-  if (runInForeground) {
-    if (verbose) {
-      logVerboseExecution({
-        projectRoot, mode: 'foreground', useApi, execContext,
-        effort, skills, resolvedModel, claudeModelAlias, explicitModel: model,
-      });
-    }
-
-    // Build claude args as array to avoid shell escaping issues with large prompts
-    const claudeArgs: string[] = [];
-    if (!process.stdin.isTTY) claudeArgs.push('--print');
-    claudeArgs.push('--dangerously-skip-permissions');
-    claudeArgs.push('--disable-slash-commands');
-    if (mcpConfigPath) claudeArgs.push('--mcp-config', mcpConfigPath);
-    if (claudeModelAlias) claudeArgs.push('--model', claudeModelAlias);
-    claudeArgs.push('--', prompt);
-
-    const agentEnv = buildAgentEnv(spawnEnv as Record<string, string>, execContext, {
-      effort, skills, includeOtel: true, ghToken: botGhToken,
-    });
-
-    return executeForeground({
-      prompt, claudeArgs, agentEnv, projectRoot: targetRepoRoot,
-      squadName, agentName, execContext, startMs, provider,
-    });
-  }
-
-  // ── Detached modes (watch + background) ──────────────────────────────
-  const timestamp = Date.now();
-  const { logFile, pidFile } = prepareLogFiles(projectRoot, squadName, agentName, timestamp);
-  const agentEnv = buildAgentEnv(spawnEnv as Record<string, string>, execContext, {
-    effort, skills, includeOtel: !runInWatch, ghToken: botGhToken,
-  });
-
-  const wrapperScript = buildDetachedShellScript({
-    projectRoot: targetRepoRoot, squadName, agentName, timestamp,
-    claudeModelAlias, escapedPrompt, logFile, pidFile,
-  });
-
-  if (runInWatch) {
-    if (verbose) {
-      logVerboseExecution({
-        projectRoot, mode: 'watch (background + tail)',
-        execContext, logFile,
-      });
-    }
-
-    return executeWatch({ projectRoot: targetRepoRoot, agentEnv, logFile, wrapperScript });
-  }
-
-  // ── Background mode ──────────────────────────────────────────────────
-  if (verbose) {
-    logVerboseExecution({
-      projectRoot, mode: 'background', useApi, execContext,
-      effort, skills, resolvedModel, claudeModelAlias,
-      explicitModel: model, logFile, mcpConfigPath,
-    });
-  }
-
-  const child = spawn('sh', ['-c', wrapperScript], {
-    cwd: targetRepoRoot,
-    detached: true,
-    stdio: 'ignore',
-    env: agentEnv,
-  });
-  child.unref();
-
-  if (verbose) {
-    writeLine(`  ${colors.dim}Monitor: tail -f ${logFile}${RESET}`);
-  }
-
-  return `Log: ${logFile}. Monitor: tail -f ${logFile}`;
-}
-
-/**
- * Execute agent with a non-Anthropic LLM CLI provider.
- *
- * Supports: google (gemini), openai (codex), mistral (vibe), xai (grok), aider, ollama
- *
- * Unlike executeWithClaude which has full session management,
- * other CLIs run in simpler non-interactive mode.
- */
-async function executeWithProvider(
-  provider: string,
-  prompt: string,
-  options: {
-    verbose?: boolean;
-    foreground?: boolean;
-    cwd?: string;
-    squadName?: string;
-    agentName?: string;
-  }
-): Promise<string> {
-  const cliConfig = getCLIConfig(provider);
-
-  if (!cliConfig) {
-    throw new Error(`Unknown provider: ${provider}. Run 'squads providers' to see available providers.`);
-  }
-
-  if (!isProviderCLIAvailable(provider)) {
-    throw new Error(`CLI '${cliConfig.command}' not found. Install: ${cliConfig.install}`);
-  }
-
-  const projectRoot = options.cwd || getProjectRoot();
-  const squadName = options.squadName || 'unknown';
-  const agentName = options.agentName || 'unknown';
-  const timestamp = Date.now();
-
-  // Build clean env: remove CLAUDECODE to allow nesting, pass squad context
-  const { CLAUDECODE: _claudeCode, ...cleanEnv } = process.env;
-  const providerEnv = {
-    ...cleanEnv,
-    SQUADS_SQUAD: squadName,
-    SQUADS_AGENT: agentName,
-    SQUADS_PROVIDER: provider,
-  };
-
-  // Create isolated worktree for this agent (same pattern as executeWithClaude)
-  const branchName = `agent/${squadName}/${agentName}-${timestamp}`;
-  const worktreePath = join(projectRoot, '..', '.worktrees', `${squadName}-${agentName}-${timestamp}`);
-  let workDir = projectRoot;
-  try {
-    mkdirSync(join(projectRoot, '..', '.worktrees'), { recursive: true });
-    execSync(`git worktree add '${worktreePath}' -b '${branchName}' HEAD`, { cwd: projectRoot, stdio: 'pipe' });
-    workDir = worktreePath;
-  } catch (e) {
-    writeLine(`  ${colors.dim}warn: worktree creation failed, using project root: ${e instanceof Error ? e.message : String(e)}${RESET}`);
-  }
-
-  // Copy .agents directory into worktree so sandboxed providers can access
-  // agent definitions, memory, and config files. Providers like Gemini restrict
-  // file reads to the workspace directory, so these must be local.
-  let effectivePrompt = prompt;
-  if (workDir !== projectRoot) {
-    const agentsDir = join(projectRoot, '.agents');
-    const targetAgentsDir = join(workDir, '.agents');
-    if (existsSync(agentsDir) && !existsSync(targetAgentsDir)) {
-      try {
-        cpSync(agentsDir, targetAgentsDir, { recursive: true });
-      } catch (e) {
-        writeLine(`  ${colors.dim}warn: .agents copy failed: ${e instanceof Error ? e.message : String(e)}${RESET}`);
-      }
-    }
-    // Rewrite absolute paths in prompt so sandboxed providers can resolve them
-    effectivePrompt = prompt.replaceAll(projectRoot, workDir);
-  }
-
-  const args = cliConfig.buildArgs(effectivePrompt);
-
-  if (options.verbose) {
-    writeLine(`  ${colors.dim}Provider: ${cliConfig.displayName}${RESET}`);
-    writeLine(`  ${colors.dim}Command: ${cliConfig.command} ${args.join(' ').slice(0, VERBOSE_COMMAND_MAX_CHARS)}...${RESET}`);
-    writeLine(`  ${colors.dim}CWD: ${workDir}${RESET}`);
-    if (workDir !== projectRoot) {
-      writeLine(`  ${colors.dim}Worktree: ${branchName}${RESET}`);
-    }
-  }
-
-  // Foreground mode: run directly in terminal
-  if (options.foreground) {
-    return new Promise((resolve, reject) => {
-      const proc = spawn(cliConfig.command, args, {
-        stdio: 'inherit',
-        cwd: workDir,
-        env: providerEnv,
-      });
-
-      proc.on('close', (code) => {
-        cleanupWorktree(workDir, projectRoot);
-        if (code === 0) {
-          resolve('Session completed');
-        } else {
-          reject(new Error(`${cliConfig.command} exited with code ${code}`));
-        }
-      });
-
-      proc.on('error', (err) => {
-        cleanupWorktree(workDir, projectRoot);
-        reject(err);
-      });
-    });
-  }
-
-  // Background mode: run detached with log file (matches executeWithClaude pattern)
-  const logDir = join(projectRoot, '.agents', 'logs', squadName);
-  const logFile = join(logDir, `${agentName}-${timestamp}.log`);
-  const pidFile = join(logDir, `${agentName}-${timestamp}.pid`);
-
-  if (!existsSync(logDir)) {
-    mkdirSync(logDir, { recursive: true });
-  }
-
-  const escapedPrompt = effectivePrompt.replace(/'/g, "'\\''");
-  const providerArgs = cliConfig.buildArgs(escapedPrompt).map(a => `'${a}'`).join(' ');
-  const cleanupCmd = workDir !== projectRoot
-    ? `; git -C '${projectRoot}' worktree remove '${workDir}' --force 2>/dev/null; git -C '${projectRoot}' branch -D '${branchName}' 2>/dev/null`
-    : '';
-  const shellScript = `cd '${workDir}' && ${cliConfig.command} ${providerArgs} > '${logFile}' 2>&1${cleanupCmd}`;
-  const wrapperScript = `echo $$ > '${pidFile}'; ${shellScript}`;
-
-  const child = spawn('sh', ['-c', wrapperScript], {
-    cwd: workDir,
-    detached: true,
-    stdio: 'ignore',
-    env: providerEnv,
-  });
-
-  child.unref();
-
-  if (options.verbose) {
-    writeLine(`  ${colors.dim}Log: ${logFile}${RESET}`);
-    writeLine(`  ${colors.dim}PID file: ${pidFile}${RESET}`);
-  }
-
-  return `Log: ${logFile}. Monitor: tail -f ${logFile}`;
-}
+// checkClaudeCliAvailable → imported from run-utils.ts
+// preflightExecutorCheck → moved to src/lib/execution-engine.ts
+
+// ExecuteWithClaudeOptions, buildAgentEnv, logVerboseExecution, resolveTargetRepoRoot,
+// createAgentWorktree, cleanupWorktree, buildDetachedShellScript, prepareLogFiles,
+// executeForeground, executeWatch, executeWithClaude, executeWithProvider
+// → moved to src/lib/execution-engine.ts
 
 export async function runSquadCommand(
   squadName: string,
